moved budget-it types to its own folder
