+++ .changeset/eleven-months-exist.md
@@ -1,0 +1,5 @@
+---
+"@inkbeard/budget-it": minor
+---
+
+Moved types to its own folder
